backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
Requests==2.31.0
yahooquery==2.3.2
django-cors-headers
pandas